refactor(desktop): split the crumb bar into its own components
A review pointed at AGENTS.md: the crumb bar had grown three
sub-components inside the parent file. PlainCrumb, AgentSwitcherCrumb
and SiblingRow now live beside it, with the shared class constants and
the switcher entry type in their own modules.

The parent drops from 282 lines to 132 and reads as what it is: the
ladder, and the decision of which crumb renders as a switcher.
