+++ apps/desktop/src/features/session/components/SessionCrumbBar/crumbClasses.ts
@@ -1,0 +1,9 @@
+export const CRUMB_BUTTON_CLASS =
+  'max-w-48 truncate rounded px-1 py-0.5 text-xs transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-[var(--color-focus-ring)]';
+
+export const CRUMB_LINK_CLASS = 'text-muted-foreground/70 hover:bg-muted/50 hover:text-foreground';
+
+export const CRUMB_LAST_CLASS = 'font-medium text-foreground';
+
+export const SIBLING_GROUP_LABEL_CLASS =
+  'px-2 pb-1 pt-2 text-3xs font-medium uppercase tracking-[0.12em] text-muted-foreground/60 first:pt-1';
